+++ slice/doc.go
@@ -71,8 +71,6 @@ func _() {
 	_ = Chunk[int]
 	_ = Compact[int]
 	_ = Contains[int]
-	_ = FromMap[int, int]
-	_ = FromMapWith[int, int, int]
 	_ = FromSet[int]
 	_ = GroupBy[int, int]
 	_ = SortBy[int, int]